feat: 增加登录接口

+++ admin-server/src/main/java/com/admin/server/dao/AppDao.java
@@ -12,7 +12,7 @@ public class AppDao {
     @Resource
     private AppMapper appMapper;
 
-    public Integer createApp(App createApp) {
+    public Long createApp(App createApp) {
         appMapper.insert(createApp);
 
         return createApp.getId();
